handle concurrent writes on same range of sectors
There is no simple way to verify concurrent IOs on same range of
sectors. We handle this by removig all traces of such IOs.

This is not the most optimal strategy, however it will suffice for
now.

Ensure there are no outstanding IOs, while switching between write and
verification mode.

Signed-off-by: Prasad Joshi <prasadjoshi.linux@gmail.com>