Add command line parameters
disk - device on which to run IOs on.

iodepth - number of outstanding IOs device

percent - percent of device to use for IO. This parameter is useful to
run IOs on limited sectors of the disk.

blocksize - List of comma separated block sizes to be used for IOs.
Along with block size, use colon (:) to specify number of times in
percent to use the block size. For example: 4096:50,8192:45,512:1 specifies
block size 4k will be used 50% of time, 8K used 45% of time, and a
sector sized IO will be used 1% of time. Block size for remaining 4% of
IOs will be selected randomly between 512 to 1M. Block size must always
be multiple of sectors.

runtime - allows user to specify how long to run tool

Signed-off-by: Prasad Joshi <prasadjoshi.linux@gmail.com>